refactor(front): セレモニー編集ルートを責務単位のコンポーネントに分割する
約1500行のadmin.ceremonies.$uid.tsxをcomponents/admin/ceremonies/配下の
ceremony-header・ceremony-form・nomination-section・nomination-table・
movie-search-panel・add-nomination-form・best-film-category・
use-awards-dataに分割し、ルートはloaderとページ骨格のみにする。
