+++ apps/front/app/components/admin/ceremonies/ensure-token.ts
@@ -1,0 +1,15 @@
+import {getAdminToken} from '@/lib/admin-fetch';
+
+export const ensureToken = () => {
+  if (typeof globalThis === 'undefined') {
+    return;
+  }
+
+  const token = getAdminToken();
+  if (!token) {
+    globalThis.location.href = '/admin/login';
+    return;
+  }
+
+  return token;
+};
