+++ apps/front/app/routes/admin.ceremonies.$uid.tsx
@@ -1,154 +1,17 @@
-import {type FormEvent, useCallback, useEffect, useMemo, useState} from 'react';
+import {useCallback, useEffect, useState} from 'react';
 import {useNavigate} from 'react-router';
-import AdminNav from '@/components/admin-nav';
-import {Button} from '@/components/ui/button';
-import {Input} from '@/components/ui/input';
-import {Textarea} from '@/components/ui/textarea';
+import {CeremonyForm} from '@/components/admin/ceremonies/ceremony-form';
+import {CeremonyHeader} from '@/components/admin/ceremonies/ceremony-header';
+import {ensureToken} from '@/components/admin/ceremonies/ensure-token';
+import {NominationSection} from '@/components/admin/ceremonies/nomination-section';
+import type {CeremonyResponse} from '@/components/admin/ceremonies/types';
+import {useAwardsData} from '@/components/admin/ceremonies/use-awards-data';
+import {adminFetch} from '@/lib/admin-fetch';
 import type {Route} from './+types/admin.ceremonies.$uid';
-import {adminFetch, getAdminToken} from '@/lib/admin-fetch';
 
 type LoaderData = {
   apiUrl: string;
   ceremonyUid: string;
-};
-
-type CeremonyNavigationItem = {
-  uid: string;
-  year: number;
-  ceremonyNumber?: number;
-};
-
-type CeremonyResponse = {
-  ceremony: {
-    uid: string;
-    organizationUid: string;
-    organizationName: string;
-    organizationCountry: string | null;
-    year: number;
-    ceremonyNumber: number | null;
-    startDate: number | null;
-    endDate: number | null;
-    location: string | null;
-    description: string | null;
-    imdbEventUrl: string | null;
-    createdAt: number;
-    updatedAt: number;
-  };
-  nominations: Array<{
-    uid: string;
-    movie: {
-      uid: string;
-      title: string;
-      year: number | null;
-    };
-    category: {
-      uid: string;
-      name: string;
-    };
-    isWinner: boolean;
-    specialMention: string | null;
-  }>;
-  navigation: {
-    previous: CeremonyNavigationItem | null;
-    next: CeremonyNavigationItem | null;
-  };
-};
-
-type AwardsOrganization = {
-  uid: string;
-  name: string;
-  country: string | null;
-  shortName: string | null;
-};
-
-type AwardsCategory = {
-  uid: string;
-  organizationUid: string;
-  name: string;
-};
-
-type AwardsData = {
-  organizations: AwardsOrganization[];
-  categories: AwardsCategory[];
-};
-
-type CeremonyFormState = {
-  organizationUid: string;
-  year: string;
-  ceremonyNumber: string;
-  startDate: string;
-  endDate: string;
-  location: string;
-  description: string;
-  imdbEventUrl: string;
-};
-
-type MovieSearchResult = {
-  uid: string;
-  title: string;
-  year: number | null;
-  imdbUrl?: string | null;
-};
-
-const emptyFormState: CeremonyFormState = {
-  organizationUid: '',
-  year: '',
-  ceremonyNumber: '',
-  startDate: '',
-  endDate: '',
-  location: '',
-  description: '',
-  imdbEventUrl: '',
-};
-
-const formatDateInput = (value: number | null) => {
-  if (!value) {
-    return '';
-  }
-
-  const date = new Date(value * 1000);
-  if (Number.isNaN(date.getTime())) {
-    return '';
-  }
-
-  return date.toISOString().slice(0, 10);
-};
-
-const formatTimestamp = (value: number) => {
-  const date = new Date(value * 1000);
-  if (Number.isNaN(date.getTime())) {
-    return '-';
-  }
-
-  return date.toLocaleString('ja-JP', {
-    year: 'numeric',
-    month: '2-digit',
-    day: '2-digit',
-    hour: '2-digit',
-    minute: '2-digit',
-  });
-};
-
-const formatNavigationLabel = (item: CeremonyNavigationItem) => {
-  if (item.ceremonyNumber && item.ceremonyNumber > 0) {
-    return `${item.year}年・第${item.ceremonyNumber}回`;
-  }
-
-  return `${item.year}年`;
-};
-
-const ensureToken = () => {
-  if (typeof globalThis === 'undefined') {
-    return;
-  }
-
-  const token = getAdminToken();
-  if (!token) {
-    globalThis.location.href = '/admin/login';
-    return;
-  }
-
-  return token;
 };
 
 export function meta() {
@@ -182,250 +45,18 @@ export default function AdminCeremonyEdit({loaderData}: Route.ComponentProps) {
   const navigate = useNavigate();
   const isNew = ceremonyUid === 'new';
 
-  const [awardsData, setAwardsData] = useState<AwardsData | undefined>();
-  const [awardsLoading, setAwardsLoading] = useState(true);
-  const [awardsError, setAwardsError] = useState<string | undefined>();
+  const {awardsData, awardsLoading, awardsError} = useAwardsData(apiUrl);
 
   const [ceremonyDetail, setCeremonyDetail] = useState<
     CeremonyResponse | undefined
   >();
   const [detailLoading, setDetailLoading] = useState(!isNew);
   const [detailError, setDetailError] = useState<string | undefined>();
-
-  const [formState, setFormState] = useState<CeremonyFormState>(emptyFormState);
-  const [saveError, setSaveError] = useState<string | undefined>();
-  const [saveSuccess, setSaveSuccess] = useState<string | undefined>();
-  const [isSaving, setIsSaving] = useState(false);
-
-  const [deleteError, setDeleteError] = useState<string | undefined>();
-  const [isDeleting, setIsDeleting] = useState(false);
-
-  const [movieSearchQuery, setMovieSearchQuery] = useState('');
-  const [movieSearchResults, setMovieSearchResults] = useState<
-    MovieSearchResult[]
-  >([]);
-  const [isSearchingMovies, setIsSearchingMovies] = useState(false);
-  const [movieSearchError, setMovieSearchError] = useState<
-    string | undefined
-  >();
-  const [selectedMovie, setSelectedMovie] = useState<MovieSearchResult>();
-
-  const [newNominationCategoryUid, setNewNominationCategoryUid] = useState('');
-  const [newNominationIsWinner, setNewNominationIsWinner] = useState(false);
-  const [newNominationSpecialMention, setNewNominationSpecialMention] =
-    useState('');
-  const [isAddingNomination, setIsAddingNomination] = useState(false);
-  const [nominationMessage, setNominationMessage] = useState<
-    string | undefined
-  >();
-  const [isSyncingFromImdb, setIsSyncingFromImdb] = useState(false);
-
-  const normalizeCategoryName = useCallback((value: string) => {
-    return value
-      .normalize('NFKC')
-      .toLowerCase()
-      .replaceAll('’', "'")
-      .replaceAll(/[（）]/g, match => (match === '（' ? '(' : ')'))
-      .replaceAll(/\s+/g, ' ')
-      .trim();
-  }, []);
-
-  const bestFilmCategory = useMemo(() => {
-    if (!awardsData) {
-      return;
-    }
-
-    const organizationUid =
-      ceremonyDetail?.ceremony.organizationUid || formState.organizationUid;
-
-    if (!organizationUid) {
-      return;
-    }
-
-    const synonyms = [
-      'best film',
-      'best picture',
-      'best motion picture',
-      'best motion picture of the year',
-      'picture of the year',
-      '最優秀作品賞',
-      '最優秀作品',
-      '作品賞',
-      '最優秀作品賞(最優秀作品)',
-      '最優秀作品賞 (最優秀作品)',
-      '作品賞(最優秀作品)',
-      '作品賞 (最優秀作品)',
-      '最優秀日本作品賞',
-      '最優秀日本映画賞',
-    ];
-
-    const normalizedSynonyms = new Set(
-      synonyms.map(synonym => normalizeCategoryName(synonym)),
-    );
-
-    let bestMatch: (typeof awardsData)['categories'][number] | undefined;
-    let bestScore = Number.POSITIVE_INFINITY;
-
-    for (const category of awardsData.categories) {
-      if (category.organizationUid !== organizationUid) {
-        continue;
-      }
-
-      const normalizedName = normalizeCategoryName(category.name);
-      if (normalizedSynonyms.has(normalizedName)) {
-        return category;
-      }
-
-      let score = Number.POSITIVE_INFINITY;
-      for (const synonym of normalizedSynonyms) {
-        if (normalizedName.includes(synonym)) {
-          score = Math.min(score, normalizedName.length - synonym.length);
-        }
-
-        if (synonym.includes(normalizedName)) {
-          score = Math.min(score, synonym.length - normalizedName.length);
-        }
-      }
-
-      if (score < bestScore) {
-        bestScore = score;
-        bestMatch = category;
-      }
-    }
-
-    return bestMatch;
-  }, [
-    awardsData,
-    ceremonyDetail?.ceremony.organizationUid,
-    formState.organizationUid,
-    normalizeCategoryName,
-  ]);
-
-  const organizationCategories = useMemo(() => {
-    if (!awardsData) {
-      return [];
-    }
-
-    const organizationUid =
-      ceremonyDetail?.ceremony.organizationUid || formState.organizationUid;
-
-    if (!organizationUid) {
-      return [];
-    }
-
-    return awardsData.categories.filter(
-      category => category.organizationUid === organizationUid,
-    );
-  }, [
-    awardsData,
-    ceremonyDetail?.ceremony.organizationUid,
-    formState.organizationUid,
-  ]);
-
-  const [syncCategoryUid, setSyncCategoryUid] = useState('');
-
-  const defaultSyncCategory = useMemo(() => {
-    if (bestFilmCategory) {
-      return bestFilmCategory;
-    }
-
-    if (organizationCategories.length === 1) {
-      return organizationCategories[0];
-    }
-
-    return;
-  }, [bestFilmCategory, organizationCategories]);
-
-  useEffect(() => {
-    if (defaultSyncCategory) {
-      setSyncCategoryUid(previous =>
-        previous === defaultSyncCategory.uid
-          ? previous
-          : defaultSyncCategory.uid,
-      );
-      return;
-    }
-
-    if (organizationCategories.length > 0) {
-      setSyncCategoryUid(previous => {
-        if (
-          previous &&
-          organizationCategories.some(category => category.uid === previous)
-        ) {
-          return previous;
-        }
-        return organizationCategories[0].uid;
-      });
-      return;
-    }
-
-    setSyncCategoryUid('');
-  }, [defaultSyncCategory, organizationCategories]);
-
-  const selectedSyncCategory = useMemo(
-    () =>
-      organizationCategories.find(category => category.uid === syncCategoryUid),
-    [organizationCategories, syncCategoryUid],
-  );
-
-  const canSyncFromImdb =
-    !isNew &&
-    Boolean(ceremonyDetail?.ceremony.imdbEventUrl) &&
-    Boolean(selectedSyncCategory);
-
-  const syncButtonTooltip = (() => {
-    if (!ceremonyDetail?.ceremony.imdbEventUrl) {
-      return 'IMDbイベントURLを設定してください。';
-    }
-
-    if (!selectedSyncCategory) {
-      return '同期対象のカテゴリを特定できませんでした。';
-    }
-
-    return;
-  })();
-
-  const isSyncButtonDisabled = isSyncingFromImdb || !canSyncFromImdb;
-
-  const fetchAwardsData = useCallback(async () => {
-    if (!ensureToken()) {
-      return;
-    }
-
-    setAwardsLoading(true);
-    setAwardsError(undefined);
-
-    try {
-      const response = await adminFetch(`${apiUrl}/admin/awards`);
-
-      if (response.status === 401) {
-        return;
-      }
-
-      if (!response.ok) {
-        throw new Error(`Failed with status ${response.status}`);
-      }
-
-      const data = (await response.json()) as {
-        organizations: AwardsOrganization[];
-        categories: AwardsCategory[];
-      };
-
-      setAwardsData({
-        organizations: data.organizations ?? [],
-        categories: data.categories ?? [],
-      });
-    } catch (error) {
-      console.error('Failed to load awards data:', error);
-      setAwardsError('主催団体・部門の取得に失敗しました。');
-    } finally {
-      setAwardsLoading(false);
-    }
-  }, [apiUrl]);
+  const [formOrganizationUid, setFormOrganizationUid] = useState('');
 
   const fetchCeremony = useCallback(
-    async (options?: {syncForm?: boolean; showSpinner?: boolean}) => {
-      const {syncForm = true, showSpinner = true} = options ?? {};
+    async (options?: {showSpinner?: boolean}) => {
+      const {showSpinner = true} = options ?? {};
 
       if (showSpinner) {
         setDetailLoading(true);
@@ -453,19 +84,6 @@ export default function AdminCeremonyEdit({loaderData}: Route.ComponentProps) {
 
         const data = (await response.json()) as CeremonyResponse;
         setCeremonyDetail(data);
-
-        if (syncForm) {
-          setFormState({
-            organizationUid: data.ceremony.organizationUid,
-            year: data.ceremony.year.toString(),
-            ceremonyNumber: data.ceremony.ceremonyNumber?.toString() ?? '',
-            startDate: formatDateInput(data.ceremony.startDate),
-            endDate: formatDateInput(data.ceremony.endDate),
-            location: data.ceremony.location ?? '',
-            description: data.ceremony.description ?? '',
-            imdbEventUrl: data.ceremony.imdbEventUrl ?? '',
-          });
-        }
       } catch (error) {
         console.error('Failed to load ceremony detail:', error);
         setDetailError('セレモニー情報の取得に失敗しました。');
@@ -480,10 +98,6 @@ export default function AdminCeremonyEdit({loaderData}: Route.ComponentProps) {
   );
 
   useEffect(() => {
-    void fetchAwardsData();
-  }, [fetchAwardsData]);
-
-  useEffect(() => {
     if (isNew) {
       setDetailLoading(false);
       setCeremonyDetail(undefined);
@@ -494,999 +108,51 @@ export default function AdminCeremonyEdit({loaderData}: Route.ComponentProps) {
       return;
     }
 
-    void fetchCeremony({syncForm: true, showSpinner: true});
+    void fetchCeremony({showSpinner: true});
   }, [fetchCeremony, isNew]);
 
-  useEffect(() => {
-    if (
-      isNew &&
-      formState.organizationUid === '' &&
-      awardsData &&
-      awardsData.organizations.length > 0
-    ) {
-      setFormState(current => ({
-        ...current,
-        organizationUid: awardsData.organizations[0]?.uid ?? '',
-      }));
-    }
-  }, [awardsData, formState.organizationUid, isNew]);
-
-  const organizationOptions = awardsData?.organizations ?? [];
-
-  const categoriesForOrganization = useMemo(() => {
-    if (awardsData === undefined) {
-      return [];
-    }
-
-    const filtered: AwardsCategory[] = awardsData.categories.filter(
-      category => category.organizationUid === formState.organizationUid,
-    );
-
-    // eslint-disable-next-line unicorn/no-array-sort
-    return filtered.sort((a, b) => a.name.localeCompare(b.name, 'ja'));
-  }, [awardsData, formState.organizationUid]);
-
-  const navigation = ceremonyDetail?.navigation;
-
-  const handleInputChange = (
-    event:
-      | FormEvent<HTMLInputElement>
-      | FormEvent<HTMLTextAreaElement>
-      | FormEvent<HTMLSelectElement>,
-  ) => {
-    const target = event.target as
-      | HTMLInputElement
-      | HTMLTextAreaElement
-      | HTMLSelectElement;
-    const {name, value} = target;
-
-    setFormState(current => ({
-      ...current,
-      [name]: value,
-    }));
-  };
-
-  const handleSave = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setSaveError(undefined);
-    setSaveSuccess(undefined);
-
-    if (!formState.organizationUid) {
-      setSaveError('主催団体を選択してください。');
-      return;
-    }
-
-    if (!formState.year) {
-      setSaveError('開催年を入力してください。');
-      return;
-    }
-
-    if (!ensureToken()) {
-      return;
-    }
-
-    setIsSaving(true);
-
-    try {
-      const payload = {
-        organizationUid: formState.organizationUid,
-        year: formState.year,
-        ceremonyNumber: formState.ceremonyNumber
-          ? Number.parseInt(formState.ceremonyNumber, 10)
-          : undefined,
-        startDate: formState.startDate || undefined,
-        endDate: formState.endDate || undefined,
-        location: formState.location || undefined,
-        description: formState.description || undefined,
-        imdbEventUrl:
-          formState.imdbEventUrl.trim() === ''
-            ? undefined
-            : formState.imdbEventUrl,
-      };
-
-      const response = await adminFetch(
-        `${apiUrl}/admin/ceremonies${
-          isNew ? '' : `/${ceremonyDetail?.ceremony.uid ?? ceremonyUid}`
-        }`,
-        {
-          method: isNew ? 'POST' : 'PUT',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(payload),
-        },
-      );
-
-      if (response.status === 401) {
-        return;
-      }
-
-      let responseBody: unknown;
-      try {
-        responseBody = await response.json();
-      } catch {
-        // ignore parse errors
-      }
-
-      if (!response.ok) {
-        const errorMessage =
-          responseBody &&
-          typeof responseBody === 'object' &&
-          'error' in responseBody &&
-          typeof (responseBody as {error?: unknown}).error === 'string'
-            ? (responseBody as {error: string}).error
-            : 'セレモニーの保存に失敗しました。';
-        throw new Error(errorMessage);
-      }
-
-      const saved = responseBody as CeremonyResponse;
-
+  const handleSaved = useCallback(
+    (saved: CeremonyResponse) => {
       setCeremonyDetail(saved);
-      setSaveSuccess('セレモニーを保存しました。');
 
       if (isNew) {
         navigate(`/admin/ceremonies/${saved.ceremony.uid}`, {replace: true});
       }
-
-      setFormState({
-        organizationUid: saved.ceremony.organizationUid,
-        year: saved.ceremony.year.toString(),
-        ceremonyNumber: saved.ceremony.ceremonyNumber?.toString() ?? '',
-        startDate: formatDateInput(saved.ceremony.startDate),
-        endDate: formatDateInput(saved.ceremony.endDate),
-        location: saved.ceremony.location ?? '',
-        description: saved.ceremony.description ?? '',
-        imdbEventUrl: saved.ceremony.imdbEventUrl ?? '',
-      });
-    } catch (error) {
-      const message =
-        error instanceof Error
-          ? error.message
-          : 'セレモニーの保存に失敗しました。';
-      setSaveError(message);
-      console.error('Save ceremony error:', error);
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const handleDeleteCeremony = async () => {
-    if (!ceremonyDetail) {
-      return;
-    }
-
-    if (typeof globalThis !== 'undefined') {
-      const confirmed = globalThis.confirm?.(
-        'このセレモニーを削除しますか？関連するノミネートも削除されます。',
-      );
-      if (!confirmed) {
-        return;
-      }
-    }
-
-    if (!ensureToken()) {
-      return;
-    }
-
-    setIsDeleting(true);
-    setDeleteError(undefined);
-
-    try {
-      const response = await adminFetch(
-        `${apiUrl}/admin/ceremonies/${ceremonyDetail.ceremony.uid}`,
-        {
-          method: 'DELETE',
-        },
-      );
-
-      if (response.status === 401) {
-        return;
-      }
-
-      if (!response.ok) {
-        const data = (await response
-          .json()
-          .catch(() => ({error: 'Unknown error'}))) as {error?: string};
-        throw new Error(data.error || 'セレモニーの削除に失敗しました。');
-      }
-
-      navigate('/admin/ceremonies');
-    } catch (error) {
-      const message =
-        error instanceof Error
-          ? error.message
-          : 'セレモニーの削除に失敗しました。';
-      setDeleteError(message);
-      console.error('Delete ceremony error:', error);
-    } finally {
-      setIsDeleting(false);
-    }
-  };
-
-  const handleSearchMovies = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setMovieSearchError(undefined);
-    setMovieSearchResults([]);
-
-    const trimmedQuery = movieSearchQuery.trim();
-    if (trimmedQuery.length < 2) {
-      setMovieSearchError('2文字以上のキーワードを入力してください。');
-      return;
-    }
-
-    if (!ensureToken()) {
-      return;
-    }
-
-    setIsSearchingMovies(true);
-
-    try {
-      const response = await adminFetch(
-        `${apiUrl}/admin/movies?limit=10&search=${encodeURIComponent(trimmedQuery)}`,
-      );
-
-      if (response.status === 401) {
-        return;
-      }
-
-      if (!response.ok) {
-        throw new Error(`Failed with status ${response.status}`);
-      }
-
-      const data = (await response.json()) as {
-        movies: Array<{
-          uid: string;
-          title: string;
-          year: number | null;
-          imdbUrl?: string | null;
-        }>;
-      };
-
-      setMovieSearchResults(data.movies ?? []);
-    } catch (error) {
-      console.error('Movie search error:', error);
-      setMovieSearchError('映画の検索に失敗しました。');
-    } finally {
-      setIsSearchingMovies(false);
-    }
-  };
-
-  const handleAddNomination = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setNominationMessage(undefined);
-
-    if (!ceremonyDetail) {
-      setNominationMessage('映画を追加する前にセレモニーを保存してください。');
-      return;
-    }
-
-    if (!selectedMovie) {
-      setNominationMessage('映画を選択してください。');
-      return;
-    }
-
-    if (!newNominationCategoryUid) {
-      setNominationMessage('部門を選択してください。');
-      return;
-    }
-
-    if (!ensureToken()) {
-      return;
-    }
-
-    setIsAddingNomination(true);
-
-    try {
-      const response = await adminFetch(
-        `${apiUrl}/admin/movies/${selectedMovie.uid}/nominations`,
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            ceremonyUid: ceremonyDetail.ceremony.uid,
-            categoryUid: newNominationCategoryUid,
-            isWinner: newNominationIsWinner,
-            specialMention:
-              newNominationSpecialMention.trim() === ''
-                ? undefined
-                : newNominationSpecialMention.trim(),
-          }),
-        },
-      );
-
-      if (response.status === 401) {
-        return;
-      }
-
-      if (!response.ok) {
-        const data = (await response
-          .json()
-          .catch(() => ({error: 'Unknown error'}))) as {error?: string};
-        throw new Error(data.error || '映画の追加に失敗しました。');
-      }
-
-      await fetchCeremony({syncForm: false, showSpinner: true});
-      setNominationMessage('映画を追加しました。');
-      setSelectedMovie(undefined);
-      setMovieSearchQuery('');
-      setMovieSearchResults([]);
-      setNewNominationCategoryUid('');
-      setNewNominationIsWinner(false);
-      setNewNominationSpecialMention('');
-    } catch (error) {
-      const message =
-        error instanceof Error ? error.message : '映画の追加に失敗しました。';
-      setNominationMessage(message);
-      console.error('Add nomination error:', error);
-    } finally {
-      setIsAddingNomination(false);
-    }
-  };
-
-  const handleRemoveNomination = async (nominationUid: string) => {
-    if (typeof globalThis !== 'undefined') {
-      const confirmed =
-        globalThis.confirm?.('この映画との紐付けを削除しますか？');
-      if (!confirmed) {
-        return;
-      }
-    }
-
-    if (!ensureToken()) {
-      return;
-    }
-
-    try {
-      const response = await adminFetch(
-        `${apiUrl}/admin/nominations/${nominationUid}`,
-        {
-          method: 'DELETE',
-        },
-      );
-
-      if (response.status === 401) {
-        return;
-      }
-
-      if (!response.ok) {
-        const data = (await response
-          .json()
-          .catch(() => ({error: 'Unknown error'}))) as {error?: string};
-        throw new Error(data.error || '紐付けの削除に失敗しました。');
-      }
-
-      await fetchCeremony({syncForm: false, showSpinner: false});
-    } catch (error) {
-      const message =
-        error instanceof Error ? error.message : '紐付けの削除に失敗しました。';
-      setNominationMessage(message);
-      console.error('Remove nomination error:', error);
-    }
-  };
-
-  const handleSyncFromImdb = async () => {
-    if (!ceremonyDetail) {
-      setNominationMessage('まずセレモニー情報を読み込んでください。');
-      return;
-    }
-
-    if (!ceremonyDetail.ceremony.imdbEventUrl) {
-      setNominationMessage('IMDbイベントURLを設定してください。');
-      return;
-    }
-
-    if (!selectedSyncCategory) {
-      setNominationMessage('同期対象のカテゴリを選択してください。');
-      return;
-    }
-
-    if (!ensureToken()) {
-      return;
-    }
-
-    setIsSyncingFromImdb(true);
-    setNominationMessage(undefined);
-
-    try {
-      const response = await adminFetch(
-        `${apiUrl}/admin/ceremonies/${ceremonyDetail.ceremony.uid}/sync-imdb`,
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({categoryUid: selectedSyncCategory.uid}),
-        },
-      );
-
-      if (response.status === 401) {
-        return;
-      }
-
-      const data = (await response.json().catch(() => ({
-        success: false,
-        error: 'Unknown error',
-      }))) as {
-        success?: boolean;
-        error?: string;
-        stats?: {
-          moviesCreated?: number;
-          nominationsInserted?: number;
-          skipped?: number;
-          imdbEntries?: number;
-          categoryName?: string;
-        };
-      };
-
-      if (!response.ok || data.success !== true) {
-        throw new Error(data.error || 'IMDbリストとの同期に失敗しました。');
-      }
-
-      await fetchCeremony({syncForm: false, showSpinner: true});
-
-      if (data.stats) {
-        const {
-          moviesCreated = 0,
-          nominationsInserted = 0,
-          skipped = 0,
-          imdbEntries = 0,
-          categoryName,
-        } = data.stats;
-
-        const detailParts = [
-          `取得 ${imdbEntries} 件`,
-          `登録 ${nominationsInserted} 件`,
-        ];
-
-        if (moviesCreated > 0) {
-          detailParts.push(`新規映画 ${moviesCreated} 件`);
-        }
-
-        if (skipped > 0) {
-          detailParts.push(`スキップ ${skipped} 件`);
-        }
-
-        setNominationMessage(
-          `IMDb${
-            categoryName
-              ? `（${categoryName}）`
-              : `（${selectedSyncCategory.name}）`
-          }のリストと同期しました（${detailParts.join(' / ')}）。`,
-        );
-      } else {
-        setNominationMessage('IMDbのリストと同期しました。');
-      }
-    } catch (error) {
-      console.error('Sync nominations from IMDb error:', error);
-      const message =
-        error instanceof Error
-          ? error.message
-          : 'IMDbリストとの同期に失敗しました。';
-      setNominationMessage(message);
-    } finally {
-      setIsSyncingFromImdb(false);
-    }
-  };
+    },
+    [isNew, navigate],
+  );
 
   return (
     <div className="min-h-screen bg-gray-100">
-      <header className="bg-white shadow">
-        <div className="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">
-          <div className="flex flex-col justify-between gap-4 md:flex-row md:items-center">
-            <div>
-              <h1 className="text-3xl font-bold text-gray-900">
-                セレモニー{isNew ? 'の新規作成' : '編集'}
-              </h1>
-              {ceremonyDetail && (
-                <p className="mt-1 text-sm text-gray-500">
-                  最終更新: {formatTimestamp(ceremonyDetail.ceremony.updatedAt)}
-                </p>
-              )}
-            </div>
-            <div className="flex flex-col gap-3 md:items-end">
-              <AdminNav />
-              {!isNew && navigation && (
-                <div className="flex justify-end gap-2">
-                  {navigation.previous ? (
-                    <Button
-                      asChild
-                      variant="outline"
-                      size="sm"
-                      className="text-gray-700">
-                      <a href={`/admin/ceremonies/${navigation.previous.uid}`}>
-                        ← {formatNavigationLabel(navigation.previous)}
-                      </a>
-                    </Button>
-                  ) : (
-                    <span className="rounded border border-gray-200 px-3 py-1 text-sm font-medium text-gray-300">
-                      ← 前へ
-                    </span>
-                  )}
-                  {navigation.next ? (
-                    <Button
-                      asChild
-                      variant="outline"
-                      size="sm"
-                      className="text-gray-700">
-                      <a href={`/admin/ceremonies/${navigation.next.uid}`}>
-                        {formatNavigationLabel(navigation.next)} →
-                      </a>
-                    </Button>
-                  ) : (
-                    <span className="rounded border border-gray-200 px-3 py-1 text-sm font-medium text-gray-300">
-                      次へ →
-                    </span>
-                  )}
-                </div>
-              )}
-              {!isNew && ceremonyDetail?.ceremony.imdbEventUrl && (
-                <Button
-                  asChild
-                  variant="outline"
-                  size="sm"
-                  className="border-blue-600 text-blue-600 hover:bg-blue-50">
-                  <a
-                    href={ceremonyDetail.ceremony.imdbEventUrl}
-                    target="_blank"
-                    rel="noreferrer">
-                    IMDbで表示
-                  </a>
-                </Button>
-              )}
-              {!isNew && (
-                <Button
-                  type="button"
-                  variant="destructive"
-                  size="sm"
-                  onClick={handleDeleteCeremony}
-                  disabled={isDeleting}
-                  className="px-4">
-                  {isDeleting ? '削除中…' : 'セレモニーを削除'}
-                </Button>
-              )}
-            </div>
-          </div>
-          {deleteError && (
-            <p className="mt-2 text-sm text-red-600">{deleteError}</p>
-          )}
-        </div>
-      </header>
+      <CeremonyHeader
+        apiUrl={apiUrl}
+        isNew={isNew}
+        ceremonyDetail={ceremonyDetail}
+      />
 
       <main className="mx-auto max-w-7xl px-4 py-8 sm:px-6 lg:px-8">
-        <section className="rounded-lg bg-white p-6 shadow">
-          <h2 className="text-lg font-semibold text-gray-900">基本情報</h2>
-          <p className="mt-1 text-sm text-gray-500">
-            主催団体や開催期間などの基本情報を編集できます。
-          </p>
+        <CeremonyForm
+          apiUrl={apiUrl}
+          ceremonyUid={ceremonyUid}
+          isNew={isNew}
+          ceremonyDetail={ceremonyDetail}
+          awardsLoading={awardsLoading}
+          awardsError={awardsError}
+          organizations={awardsData?.organizations ?? []}
+          onSaved={handleSaved}
+          onOrganizationUidChange={setFormOrganizationUid}
+        />
 
-          <form className="mt-6 space-y-5" onSubmit={handleSave}>
-            {awardsLoading ? (
-              <div className="rounded bg-gray-50 px-4 py-3 text-sm text-gray-500">
-                主催団体を読み込み中です…
-              </div>
-            ) : awardsError ? (
-              <div className="rounded bg-red-50 px-4 py-3 text-sm text-red-700">
-                {awardsError}
-              </div>
-            ) : (
-              <div className="grid gap-4 md:grid-cols-2">
-                <label className="flex flex-col text-sm font-medium text-gray-700">
-                  主催団体
-                  <select
-                    name="organizationUid"
-                    value={formState.organizationUid}
-                    onChange={handleInputChange}
-                    className="mt-1 rounded border border-gray-300 px-3 py-2 shadow-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
-                    required>
-                    <option value="">選択してください</option>
-                    {organizationOptions.map(organization => (
-                      <option key={organization.uid} value={organization.uid}>
-                        {organization.name}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-
-                <label className="flex flex-col text-sm font-medium text-gray-700">
-                  開催年
-                  <Input
-                    type="number"
-                    name="year"
-                    value={formState.year}
-                    onChange={handleInputChange}
-                    placeholder="2025"
-                    className="mt-1"
-                    required
-                  />
-                </label>
-
-                <label className="flex flex-col text-sm font-medium text-gray-700">
-                  回数
-                  <Input
-                    type="number"
-                    name="ceremonyNumber"
-                    value={formState.ceremonyNumber}
-                    onChange={handleInputChange}
-                    placeholder="例: 96"
-                    className="mt-1"
-                    min={1}
-                  />
-                </label>
-
-                <div className="grid gap-4 md:grid-cols-2">
-                  <label className="flex flex-col text-sm font-medium text-gray-700">
-                    開始日
-                    <Input
-                      type="date"
-                      name="startDate"
-                      value={formState.startDate}
-                      onChange={handleInputChange}
-                      className="mt-1"
-                    />
-                  </label>
-
-                  <label className="flex flex-col text-sm font-medium text-gray-700">
-                    終了日
-                    <Input
-                      type="date"
-                      name="endDate"
-                      value={formState.endDate}
-                      onChange={handleInputChange}
-                      className="mt-1"
-                    />
-                  </label>
-                </div>
-
-                <label className="md:col-span-2 flex flex-col text-sm font-medium text-gray-700">
-                  開催場所
-                  <Input
-                    type="text"
-                    name="location"
-                    value={formState.location}
-                    onChange={handleInputChange}
-                    placeholder="例: ロサンゼルス"
-                    className="mt-1"
-                  />
-                </label>
-
-                <label className="md:col-span-2 flex flex-col text-sm font-medium text-gray-700">
-                  IMDbイベントURL
-                  <Input
-                    type="url"
-                    name="imdbEventUrl"
-                    value={formState.imdbEventUrl}
-                    onChange={handleInputChange}
-                    placeholder="https://www.imdb.com/event/ev0000372/1978/1"
-                    className="mt-1"
-                  />
-                  <span className="mt-1 text-xs text-gray-500">
-                    IMDb のイベントページへの完全な URL
-                    を入力してください（任意）。
-                  </span>
-                </label>
-
-                <label className="md:col-span-2 flex flex-col text-sm font-medium text-gray-700">
-                  説明
-                  <Textarea
-                    name="description"
-                    value={formState.description}
-                    onChange={handleInputChange}
-                    rows={4}
-                    placeholder="補足情報があれば記入してください"
-                    className="mt-1"
-                  />
-                </label>
-              </div>
-            )}
-
-            {saveError && (
-              <div className="rounded bg-red-50 px-4 py-3 text-sm text-red-700">
-                {saveError}
-              </div>
-            )}
-
-            {saveSuccess && (
-              <div className="rounded bg-green-50 px-4 py-3 text-sm text-green-700">
-                {saveSuccess}
-              </div>
-            )}
-
-            <div className="flex justify-end gap-3">
-              <Button
-                type="submit"
-                disabled={isSaving || awardsLoading}
-                className="bg-blue-600 text-white hover:bg-blue-500 disabled:cursor-not-allowed disabled:opacity-60">
-                {isSaving ? '保存中…' : '保存する'}
-              </Button>
-            </div>
-          </form>
-        </section>
-
-        <section className="mt-8 rounded-lg bg-white p-6 shadow">
-          <div className="flex flex-col justify-between gap-3 md:flex-row md:items-center">
-            <div>
-              <h2 className="text-lg font-semibold text-gray-900">
-                紐付いている映画
-              </h2>
-              <p className="mt-1 text-sm text-gray-500">
-                ノミネート・受賞作品を追加・削除できます。
-              </p>
-            </div>
-            <div className="flex flex-wrap items-center gap-3">
-              {ceremonyDetail && (
-                <span className="rounded-full bg-gray-100 px-3 py-1 text-sm text-gray-600">
-                  {ceremonyDetail.nominations.length} 件
-                </span>
-              )}
-              {!isNew && (
-                <Button
-                  type="button"
-                  onClick={handleSyncFromImdb}
-                  disabled={isSyncButtonDisabled}
-                  title={syncButtonTooltip}
-                  className="bg-indigo-600 text-white hover:bg-indigo-500 disabled:cursor-not-allowed disabled:opacity-60">
-                  {isSyncingFromImdb ? '同期中…' : 'IMDbリストと同期'}
-                </Button>
-              )}
-              {!isNew && organizationCategories.length > 1 && (
-                <label className="flex items-center gap-2 text-sm text-gray-700">
-                  <span>対象部門</span>
-                  <select
-                    value={syncCategoryUid}
-                    onChange={event => setSyncCategoryUid(event.target.value)}
-                    className="rounded border border-gray-300 px-2 py-1 text-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500">
-                    {organizationCategories.map(category => (
-                      <option key={category.uid} value={category.uid}>
-                        {category.name}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-              )}
-            </div>
-          </div>
-
-          {detailLoading ? (
-            <div className="mt-6 rounded bg-gray-50 px-4 py-3 text-sm text-gray-500">
-              ノミネート情報を読み込み中です…
-            </div>
-          ) : detailError ? (
-            <div className="mt-6 rounded bg-red-50 px-4 py-3 text-sm text-red-700">
-              {detailError}
-            </div>
-          ) : ceremonyDetail ? (
-            <div className="mt-6 space-y-6">
-              {ceremonyDetail.nominations.length === 0 ? (
-                <div className="rounded border border-dashed border-gray-300 px-4 py-6 text-center text-sm text-gray-500">
-                  登録されている映画はありません。
-                </div>
-              ) : (
-                <div className="overflow-x-auto">
-                  <table className="min-w-full divide-y divide-gray-200">
-                    <thead className="bg-gray-50">
-                      <tr>
-                        <th className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-gray-500">
-                          映画
-                        </th>
-                        <th className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-gray-500">
-                          部門
-                        </th>
-                        <th className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-gray-500">
-                          受賞
-                        </th>
-                        <th className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-gray-500">
-                          特記事項
-                        </th>
-                        <th className="px-4 py-3 text-right text-xs font-semibold uppercase tracking-wider text-gray-500">
-                          操作
-                        </th>
-                      </tr>
-                    </thead>
-                    <tbody className="divide-y divide-gray-200 bg-white">
-                      {ceremonyDetail.nominations.map(nomination => (
-                        <tr key={nomination.uid}>
-                          <td className="px-4 py-3 text-sm text-gray-900">
-                            <div className="font-medium text-gray-900">
-                              <a
-                                href={`/admin/movies/${nomination.movie.uid}`}
-                                className="text-blue-600 hover:underline">
-                                {nomination.movie.title}
-                              </a>
-                            </div>
-                            <div className="text-xs text-gray-500">
-                              UID: {nomination.movie.uid}
-                              {nomination.movie.year
-                                ? ` / ${nomination.movie.year}年`
-                                : ''}
-                            </div>
-                          </td>
-                          <td className="px-4 py-3 text-sm text-gray-900">
-                            {nomination.category.name}
-                          </td>
-                          <td className="px-4 py-3 text-sm text-gray-900">
-                            {nomination.isWinner ? '受賞' : 'ノミネート'}
-                          </td>
-                          <td className="px-4 py-3 text-sm text-gray-900">
-                            {nomination.specialMention ?? '-'}
-                          </td>
-                          <td className="px-4 py-3 text-right">
-                            <Button
-                              type="button"
-                              variant="outline"
-                              size="sm"
-                              className="border-red-600 text-red-600 hover:bg-red-50"
-                              onClick={() =>
-                                handleRemoveNomination(nomination.uid)
-                              }>
-                              削除
-                            </Button>
-                          </td>
-                        </tr>
-                      ))}
-                    </tbody>
-                  </table>
-                </div>
-              )}
-
-              <div className="rounded-lg border border-gray-200 p-4">
-                <h3 className="text-md font-semibold text-gray-900">
-                  映画を追加
-                </h3>
-                <p className="mt-1 text-sm text-gray-500">
-                  映画を検索し、部門を選択して追加します。
-                </p>
-
-                <div className="mt-4 space-y-6">
-                  <form
-                    className="flex flex-col gap-3 md:flex-row md:items-end"
-                    onSubmit={handleSearchMovies}>
-                    <label className="flex flex-1 flex-col text-sm font-medium text-gray-700">
-                      キーワード検索
-                      <Input
-                        type="search"
-                        value={movieSearchQuery}
-                        onChange={event =>
-                          setMovieSearchQuery(event.target.value)
-                        }
-                        placeholder="作品名など"
-                        className="mt-1"
-                      />
-                    </label>
-                    <Button
-                      type="submit"
-                      className="bg-blue-600 text-white hover:bg-blue-500 disabled:cursor-not-allowed disabled:opacity-60"
-                      disabled={isSearchingMovies}>
-                      {isSearchingMovies ? '検索中…' : '検索'}
-                    </Button>
-                  </form>
-
-                  {movieSearchError && (
-                    <div className="rounded bg-red-50 px-4 py-3 text-sm text-red-700">
-                      {movieSearchError}
-                    </div>
-                  )}
-
-                  {selectedMovie && (
-                    <div className="rounded border border-blue-200 bg-blue-50 px-4 py-3 text-sm text-blue-700">
-                      選択中: {selectedMovie.title}
-                      {selectedMovie.year ? `（${selectedMovie.year}年）` : ''}
-                      <Button
-                        type="button"
-                        variant="link"
-                        size="sm"
-                        className="ml-2 text-blue-700 hover:text-blue-900"
-                        onClick={() => setSelectedMovie(undefined)}>
-                        解除
-                      </Button>
-                    </div>
-                  )}
-
-                  {movieSearchResults.length > 0 && (
-                    <div className="rounded border border-gray-200">
-                      <ul className="divide-y divide-gray-200">
-                        {movieSearchResults.map(result => (
-                          <li
-                            key={result.uid}
-                            className="flex items-center justify-between px-4 py-3 text-sm">
-                            <div>
-                              <div className="font-medium text-gray-900">
-                                {result.title}
-                              </div>
-                              <div className="text-xs text-gray-500">
-                                UID: {result.uid}
-                                {result.year ? ` / ${result.year}年` : ''}
-                              </div>
-                            </div>
-                            <Button
-                              type="button"
-                              variant="outline"
-                              size="sm"
-                              className="border-blue-600 text-blue-600 hover:bg-blue-50"
-                              onClick={() => {
-                                setSelectedMovie(result);
-                                setMovieSearchResults([]);
-                              }}>
-                              選択
-                            </Button>
-                          </li>
-                        ))}
-                      </ul>
-                    </div>
-                  )}
-
-                  <form className="space-y-4" onSubmit={handleAddNomination}>
-                    <div className="grid gap-4 md:grid-cols-2">
-                      <label className="flex flex-col text-sm font-medium text-gray-700">
-                        部門
-                        <select
-                          value={newNominationCategoryUid}
-                          onChange={event =>
-                            setNewNominationCategoryUid(event.target.value)
-                          }
-                          className="mt-1 rounded border border-gray-300 px-3 py-2 shadow-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
-                          required
-                          disabled={categoriesForOrganization.length === 0}>
-                          <option value="">選択してください</option>
-                          {categoriesForOrganization.map(category => (
-                            <option key={category.uid} value={category.uid}>
-                              {category.name}
-                            </option>
-                          ))}
-                        </select>
-                      </label>
-
-                      <label className="flex items-center gap-2 text-sm font-medium text-gray-700">
-                        <input
-                          type="checkbox"
-                          checked={newNominationIsWinner}
-                          onChange={event =>
-                            setNewNominationIsWinner(event.target.checked)
-                          }
-                          className="h-4 w-4 rounded border-gray-300 text-blue-600 focus:ring-blue-500"
-                        />
-                        受賞として登録
-                      </label>
-                    </div>
-
-                    <label className="flex flex-col text-sm font-medium text-gray-700">
-                      特記事項
-                      <Input
-                        type="text"
-                        value={newNominationSpecialMention}
-                        onChange={event =>
-                          setNewNominationSpecialMention(event.target.value)
-                        }
-                        placeholder="コメント等（任意）"
-                        className="mt-1"
-                      />
-                    </label>
-
-                    {nominationMessage && (
-                      <div className="rounded bg-blue-50 px-4 py-3 text-sm text-blue-700">
-                        {nominationMessage}
-                      </div>
-                    )}
-
-                    <div className="flex justify-end">
-                      <Button
-                        type="submit"
-                        disabled={!selectedMovie || isAddingNomination}
-                        className="bg-green-600 text-white hover:bg-green-500 disabled:cursor-not-allowed disabled:opacity-60">
-                        {isAddingNomination ? '追加中…' : '映画を追加'}
-                      </Button>
-                    </div>
-                  </form>
-                </div>
-              </div>
-            </div>
-          ) : (
-            <div className="mt-6 rounded border border-dashed border-gray-300 px-4 py-6 text-center text-sm text-gray-500">
-              まずセレモニー情報を保存してください。
-            </div>
-          )}
-        </section>
+        <NominationSection
+          apiUrl={apiUrl}
+          isNew={isNew}
+          ceremonyDetail={ceremonyDetail}
+          detailLoading={detailLoading}
+          detailError={detailError}
+          awardsData={awardsData}
+          formOrganizationUid={formOrganizationUid}
+          refetchCeremony={fetchCeremony}
+        />
       </main>
     </div>
   );
